Add files via upload
These files consist of the server.js, the .env file to connect to the database server. the browser.py to connect the html to the database. Controllers are used for functions, Routes give directions from the html to the controllers, models create the database in mongoDB. To use these files run the server.js file after changing the URL of the .env file to be the URL of your mongoDB URL.